chore(gitignore): update .gitignore to include common Python, LaTeX, and system files
Expand the .gitignore file to cover a broader range of files
and directories typically generated in Python projects, such as
byte-compiled files, virtual environments, and distribution
artifacts. Also, include common system files for macOS and Windows,
as well as Jupyter Notebook checkpoints. This helps in keeping the
repository clean by ignoring unnecessary files.

docs: remove outdated readme.md and add new README.md for photos

Remove the outdated readme.md file from the 2023/code directory
as it is no longer relevant. Add a new README.md in the 2023/photos
directory to provide a visual reference with images related to
the project. This enhances the documentation by focusing on
current and relevant information.